refactor: expense -> shared 로 변경
