--- snapsplit/src/app/(no-gnb)/trip/[tripId]/budget/expense/[mode]/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/features/trip/[tripId]/budget/expense/ExpensPage';
