+++ snapsplit/src/app/(no-gnb)/trip/[tripId]/budget/expense/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/trip/[tripId]/budget/expense/ExpensePage";
