+++ snapsplit/src/app/(no-gnb)/trip/[tripId]/budget/shared/[mode]/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/features/trip/[tripId]/budget/shared/[mode]/SharedPage';
