--- snapsplit/src/app/(no-gnb)/trip/[tripId]/budget/spend/page.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/features/trip/[tripId]/budget/spend/SpendPage";
