--- snapsplit/src/features/trip/[tripId]/budget/expense/ExpensPage.tsx
@@ -1,7 +1,0 @@
-import ExpenseForm from '@/features/trip/[tripId]/budget/expense/_components/ExpenseForm';
-
-const ExpensePage = () => {
-  return <ExpenseForm />;
-};
-
-export default ExpensePage;
