+++ snapsplit/src/features/trip/[tripId]/budget/expense/_components/spendForm/DescriptionSection.tsx
