+++ snapsplit/src/features/trip/[tripId]/budget/shared/[mode]/SharedPage.tsx
@@ -1,0 +1,7 @@
+import SharedForm from '@/features/trip/[tripId]/budget/shared/[mode]/_components/SharedForm';
+
+const SharedPage = () => {
+  return <SharedForm />;
+};
+
+export default SharedPage;
