+++ snapsplit/src/features/trip/[tripId]/budget/shared/[mode]/_components/CategorySection.tsx
